site status management for movies
