+++ src/common/Pagination/index.js
@@ -1,4 +1,7 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
+import { useDispatch } from "react-redux";
+import { fetchGenres, fetchMovies } from "../../features/movies/moviesSlice";
+import { fetchPeople } from "../../features/people/peopleSlice";
 import {
   StyledPagination,
   ButtonsWrapper,
@@ -13,6 +16,13 @@ import {
 
 const Pagination = () => {
   const [currentPage, setCurrentPage] = useState(1);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchGenres());
+    dispatch(fetchMovies());
+    dispatch(fetchPeople());
+  }, []);
 
   const onGoToFirst = () => {
     if (currentPage !== 1) {
