--- src/features/movies/PopularMoviesPage/styled.js
@@ -1,1 +1,0 @@
-
