+++ src/features/movies/moviesSaga.js
@@ -13,7 +13,10 @@ function* fetchGenresHandler() {
   try {
     const genres = yield call(getGenres);
     yield put(setGenres(genres));
+    yield delay(1000);
+    yield put(setFetchingToSucces());
   } catch (error) {
+    yield delay(1000);
     yield put(setFetchingToFail());
   }
 }
@@ -22,8 +25,6 @@ function* fetchMoviesHandler() {
   try {
     const movies = yield call(getPopularData, "movie");
     yield put(setMovies(movies));
-    yield delay(1000);
-    yield put(setFetchingToSucces());
   } catch (error) {
     yield delay(1000);
     yield put(setFetchingToFail());
