+++ src/features/people/PopularPeoplePage/index.js
@@ -1,24 +1,17 @@
-import { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { selectPageStatus, selectPeople, fetchPeople } from "../peopleSlice";
+import { useSelector } from "react-redux";
+import { selectFetchingStatus } from "../peopleSlice";
 import { Loading } from "../../../common/Loading";
 import ErrorPage from "../../../common/ErrorPage";
 import PopularPeople from "./PopularPeople";
 
 const PopularPeoplePage = () => {
-  const pageStatus = useSelector(selectPageStatus);
-  const people = useSelector(selectPeople);
-  const dispatch = useDispatch();
+  const fetchingStatus = useSelector(selectFetchingStatus);
 
-  useEffect(() => {
-    dispatch(fetchPeople());
-  }, []);
-  
   return {
     loading: <Loading />,
     success: <PopularPeople />,
     fail: <ErrorPage />,
-  }[pageStatus];
+  }[fetchingStatus];
 };
 
 export default PopularPeoplePage;
