+++ src/features/people/peopleSlice.js
@@ -4,24 +4,24 @@ const peopleSlice = createSlice({
     name: 'people',
     initialState: {
         people: [],
-        loadingStatus: "loading",
+        fetchingStatus: "loading",
     },
     reducers: {
         setPeople: (state, { payload: people }) => {
             state.people = people
         },
-        setLoadingToSucces: state => {
-            state.loadingStatus = "success";
+        setFetchingToSucces: state => {
+            state.fetchingStatus = "success";
         },
-        setLoadingToFail: state => {
-            state.loadingStatus = "fail";
+        setFetchingToFail: state => {
+            state.fetchingStatus = "fail";
         },
         fetchPeople: () => { },
     },
 });
 
-export const { setPeople, setLoadingToSucces, setLoadingToFail, fetchPeople } = peopleSlice.actions;
+export const { setPeople, setFetchingToSucces, setFetchingToFail, fetchPeople } = peopleSlice.actions;
 export const selectPeopleState = state => state.people;
 export const selectPeople = state => selectPeopleState(state).people;
-export const selectPageStatus = state => selectPeopleState(state).loadingStatus;
+export const selectFetchingStatus = state => selectPeopleState(state).fetchingStatus;
 export default peopleSlice.reducer;
